refcandidates: small refactor
